+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-# [Unreleased]
+## [7.2.3] 2025-03-11
 
 ### Fixed
 
-- [TD-7145]: 
+- [TD-7145]:
   - Handle empty fields in xlsx bulk upload for notes in status other than `draft`.
   - Put empty fields in row tail.
   - Merge content before parsing.